# backend/tests/snapshots/openapi.json
{
  "components": {
    "schemas": {
      "AbandonedReason": {
        "enum": [
          "too_hard",
          "too_easy",
          "no_time",
          "lost_interest",
          "blocked",
          "other"
        ],
        "title": "AbandonedReason",
        "type": "string"
      },
      "ActivePlanResponse": {
        "additionalProperties": false,
        "properties": {
          "adjustment_reason": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Adjustment Reason"
          },
          "adopted_at": {
            "anyOf": [
              {
                "format": "date-time",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Adopted At"
          },
          "companion_message": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Companion Message"
          },
          "created_at": {
            "format": "date-time",
            "title": "Created At",
            "type": "string"
          },
          "horizon_end": {
            "format": "date",
            "title": "Horizon End",
            "type": "string"
          },
          "horizon_start": {
            "format": "date",
            "title": "Horizon Start",
            "type": "string"
          },
          "overall_direction": {
            "title": "Overall Direction",
            "type": "string"
          },
          "plan_date": {
            "format": "date",
            "title": "Plan Date",
            "type": "string"
          },
          "plan_id": {
            "format": "uuid",
            "title": "Plan Id",
            "type": "string"
          },
          "rationale": {
            "title": "Rationale",
            "type": "string"
          },
          "sources": {
            "items": {
              "$ref": "#/components/schemas/PlanSourceResponse"
            },
            "title": "Sources",
            "type": "array"
          },
          "status": {
            "$ref": "#/components/schemas/PlanStatus"
          },
          "summary": {
            "title": "Summary",
            "type": "string"
          },
          "tasks": {
            "items": {
              "$ref": "#/components/schemas/TaskResponse"
            },
            "title": "Tasks",
            "type": "array"
          },
          "version": {
            "title": "Version",
            "type": "integer"
          },
          "weekly_focus": {
            "items": {
              "$ref": "#/components/schemas/WeeklyFocusCandidate"
            },
            "title": "Weekly Focus",
            "type": "array"
          }
        },
        "required": [
          "plan_id",
          "status",
          "plan_date",
          "horizon_start",
          "horizon_end",
          "overall_direction",
          "weekly_focus",
          "summary",
          "rationale",
          "adjustment_reason",
          "tasks",
          "companion_message",
          "version",
          "adopted_at",
          "created_at"
        ],
        "title": "ActivePlanResponse",
        "type": "object"
      },
...
      "ErrorBody": {
        "additionalProperties": false,
        "description": "Details contained in the top-level error envelope.",
        "properties": {
          "code": {
            "title": "Code",
            "type": "string"
          },
          "details": {
            "additionalProperties": {
              "$ref": "#/components/schemas/JsonValue"
            },
            "title": "Details",
            "type": "object"
          },
          "message": {
            "title": "Message",
            "type": "string"
          },
          "request_id": {
            "title": "Request Id",
            "type": "string"
          }
        },
        "required": [
          "code",
          "message",
          "request_id"
        ],
        "title": "ErrorBody",
        "type": "object"
      },
      "ErrorResponse": {
        "additionalProperties": false,
        "description": "Stable API error envelope.",
        "properties": {
          "error": {
            "$ref": "#/components/schemas/ErrorBody"
          }
        },
        "required": [
          "error"
        ],
        "title": "ErrorResponse",
        "type": "object"
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "title": "Detail",
            "type": "array"
          }
        },
        "title": "HTTPValidationError",
        "type": "object"
      },
...
      "JsonValue": {},
...
      "PlanListResponse": {
        "additionalProperties": false,
        "properties": {
          "items": {
            "items": {
              "$ref": "#/components/schemas/ActivePlanResponse"
            },
            "title": "Items",
            "type": "array"
          },
          "next_cursor": {
            "anyOf": [
              {
                "format": "uuid",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Next Cursor"
          }
        },
        "required": [
          "items",
          "next_cursor"
        ],
        "title": "PlanListResponse",
        "type": "object"
      },
...
      "PlanSourceResponse": {
        "additionalProperties": false,
        "properties": {
          "available": {
            "title": "Available",
            "type": "boolean"
          },
          "id": {
            "format": "uuid",
            "title": "Id",
            "type": "string"
          },
          "kind": {
            "enum": [
              "memory",
              "experience_atom",
              "search_source"
            ],
            "title": "Kind",
            "type": "string"
          },
          "reliability": {
            "anyOf": [
              {
                "maximum": 1.0,
                "minimum": 0.0,
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Reliability"
          },
          "snippet": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Snippet"
          },
          "title": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Title"
          },
          "url": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Url"
          }
        },
        "required": [
          "kind",
          "id",
          "available"
        ],
        "title": "PlanSourceResponse",
        "type": "object"
      },
      "PlanSourcesResponse": {
        "additionalProperties": false,
        "properties": {
          "items": {
            "items": {
              "$ref": "#/components/schemas/PlanSourceResponse"
            },
            "title": "Items",
            "type": "array"
          }
        },
        "required": [
          "items"
        ],
        "title": "PlanSourcesResponse",
        "type": "object"
      },
      "PlanStatus": {
        "enum": [
          "generated",
          "active",
          "completed",
          "archived"
        ],
        "title": "PlanStatus",
        "type": "string"
      },
...
      "TaskListResponse": {
        "additionalProperties": false,
        "properties": {
          "items": {
            "items": {
              "$ref": "#/components/schemas/TaskResponse"
            },
            "title": "Items",
            "type": "array"
          }
        },
        "required": [
          "items"
        ],
        "title": "TaskListResponse",
        "type": "object"
      },
      "TaskResponse": {
        "additionalProperties": false,
        "properties": {
          "abandoned_at": {
            "anyOf": [
              {
                "format": "date-time",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Abandoned At"
          },
          "abandoned_reason": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/AbandonedReason"
              },
              {
                "type": "null"
              }
            ]
          },
          "abandoned_reason_text": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Abandoned Reason Text"
          },
          "actual_minutes": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Actual Minutes"
          },
          "completed_at": {
            "anyOf": [
              {
                "format": "date-time",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Completed At"
          },
          "created_at": {
            "format": "date-time",
            "title": "Created At",
            "type": "string"
          },
          "deliverable": {
            "title": "Deliverable",
            "type": "string"
          },
          "estimated_minutes": {
            "title": "Estimated Minutes",
            "type": "integer"
          },
          "order_index": {
            "minimum": 0.0,
            "title": "Order Index",
            "type": "integer"
          },
          "plan_id": {
            "format": "uuid",
            "title": "Plan Id",
            "type": "string"
          },
          "rationale": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Rationale"
          },
          "scheduled_date": {
            "format": "date",
            "title": "Scheduled Date",
            "type": "string"
          },
          "started_at": {
            "anyOf": [
              {
                "format": "date-time",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Started At"
          },
          "starter_action": {
            "title": "Starter Action",
            "type": "string"
          },
          "state": {
            "$ref": "#/components/schemas/TaskStatus"
          },
          "task_id": {
            "format": "uuid",
            "title": "Task Id",
            "type": "string"
          },
          "task_type": {
            "$ref": "#/components/schemas/TaskType"
          },
          "title": {
            "title": "Title",
            "type": "string"
          },
          "version": {
            "title": "Version",
            "type": "integer"
          }
        },
        "required": [
          "task_id",
          "plan_id",
          "title",
          "task_type",
          "scheduled_date",
          "order_index",
          "state",
          "starter_action",
          "deliverable",
          "rationale",
          "estimated_minutes",
          "actual_minutes",
          "abandoned_reason",
          "abandoned_reason_text",
          "version",
          "started_at",
          "completed_at",
          "abandoned_at",
          "created_at"
        ],
        "title": "TaskResponse",
        "type": "object"
      },
      "TaskStatus": {
        "enum": [
          "pending",
          "in_progress",
          "completed",
          "abandoned",
          "expired"
        ],
        "title": "TaskStatus",
        "type": "string"
      },
      "TaskType": {
        "enum": [
          "learning",
          "project",
          "interview",
          "application",
          "resume",
          "other"
        ],
        "title": "TaskType",
        "type": "string"
      },
      "TaskUpdateRequest": {
        "additionalProperties": false,
        "properties": {
          "abandoned_reason": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/AbandonedReason"
              },
              {
                "type": "null"
              }
            ]
          },
          "abandoned_reason_text": {
            "anyOf": [
              {
                "maxLength": 200,
                "minLength": 1,
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Abandoned Reason Text"
          },
          "actual_minutes": {
            "anyOf": [
              {
                "maximum": 1440.0,
                "minimum": 0.0,
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Actual Minutes"
          },
          "state": {
            "enum": [
              "in_progress",
              "completed",
              "abandoned"
            ],
            "title": "State",
            "type": "string"
          },
          "version": {
            "minimum": 1.0,
            "title": "Version",
            "type": "integer"
          }
        },
        "required": [
          "state",
          "version"
        ],
        "title": "TaskUpdateRequest",
        "type": "object"
      },
      "TaskUpdateResponse": {
        "additionalProperties": false,
        "properties": {
          "companion_message": {
            "title": "Companion Message",
            "type": "string"
          },
          "plan_status": {
            "$ref": "#/components/schemas/PlanStatus"
          },
          "task": {
            "$ref": "#/components/schemas/TaskResponse"
          }
        },
        "required": [
          "task",
          "plan_status",
          "companion_message"
        ],
        "title": "TaskUpdateResponse",
        "type": "object"
      },
...
      "ValidationError": {
        "properties": {
          "ctx": {
            "title": "Context",
            "type": "object"
          },
          "input": {
            "title": "Input"
          },
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "title": "Location",
            "type": "array"
          },
          "msg": {
            "title": "Message",
            "type": "string"
          },
          "type": {
            "title": "Error Type",
            "type": "string"
          }
        },
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError",
        "type": "object"
      },
      "WeeklyFocusCandidate": {
        "additionalProperties": false,
        "properties": {
          "focus": {
            "maxLength": 160,
            "minLength": 1,
            "title": "Focus",
            "type": "string"
          },
          "success_signal": {
            "maxLength": 200,
            "minLength": 1,
            "title": "Success Signal",
            "type": "string"
          },
          "week_index": {
            "maximum": 8.0,
            "minimum": 1.0,
            "title": "Week Index",
            "type": "integer"
          }
        },
        "required": [
          "week_index",
          "focus",
          "success_signal"
        ],
        "title": "WeeklyFocusCandidate",
        "type": "object"
      }
    },
    "securitySchemes": {
      "HTTPBearer": {
        "scheme": "bearer",
        "type": "http"
      }
    }
  },
  "info": {
    "title": "Career Planning Buddy",
    "version": "0.1.0"
  },
  "openapi": "3.1.0",
  "paths": {
...
    "/api/v1/plans": {
      "get": {
        "operationId": "list_plans_api_v1_plans_get",
        "parameters": [
          {
            "in": "query",
            "name": "status",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "$ref": "#/components/schemas/PlanStatus"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Status"
            }
          },
          {
            "in": "query",
            "name": "date_from",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "format": "date",
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Date From"
            }
          },
          {
            "in": "query",
            "name": "date_to",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "format": "date",
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Date To"
            }
          },
          {
            "in": "query",
            "name": "cursor",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "format": "uuid",
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Cursor"
            }
          },
          {
            "in": "query",
            "name": "limit",
            "required": false,
            "schema": {
              "default": 50,
              "maximum": 100,
              "minimum": 1,
              "title": "Limit",
              "type": "integer"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PlanListResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "401": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Unauthorized"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "summary": "List Plans",
        "tags": [
          "plans"
        ]
      }
    },
    "/api/v1/plans/active": {
      "get": {
        "operationId": "get_active_plan_api_v1_plans_active_get",
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ActivePlanResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "401": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Unauthorized"
          },
          "404": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Not Found"
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "summary": "Get Active Plan",
        "tags": [
          "plans"
        ]
      }
    },
    "/api/v1/plans/{plan_id}": {
      "get": {
        "operationId": "get_plan_api_v1_plans__plan_id__get",
        "parameters": [
          {
            "in": "path",
            "name": "plan_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Plan Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ActivePlanResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "401": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Unauthorized"
          },
          "404": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Not Found"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "summary": "Get Plan",
        "tags": [
          "plans"
        ]
      }
    },
    "/api/v1/plans/{plan_id}/sources": {
      "get": {
        "operationId": "get_plan_sources_api_v1_plans__plan_id__sources_get",
        "parameters": [
          {
            "in": "path",
            "name": "plan_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Plan Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PlanSourcesResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "401": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Unauthorized"
          },
          "404": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Not Found"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "summary": "Get Plan Sources",
        "tags": [
          "plans"
        ]
      }
    },
...
    "/api/v1/tasks": {
      "get": {
        "operationId": "list_tasks_api_v1_tasks_get",
        "parameters": [
          {
            "in": "query",
            "name": "date",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "format": "date",
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Date"
            }
          },
          {
            "in": "query",
            "name": "state",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "$ref": "#/components/schemas/TaskStatus"
                },
                {
                  "type": "null"
                }
              ],
              "title": "State"
            }
          },
          {
            "in": "query",
            "name": "plan_id",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "format": "uuid",
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Plan Id"
            }
          },
          {
            "in": "query",
            "name": "limit",
            "required": false,
            "schema": {
              "default": 50,
              "maximum": 100,
              "minimum": 1,
              "title": "Limit",
              "type": "integer"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/TaskListResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "401": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Unauthorized"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "summary": "List Tasks",
        "tags": [
          "tasks"
        ]
      }
    },
    "/api/v1/tasks/{task_id}": {
      "patch": {
        "operationId": "update_task_api_v1_tasks__task_id__patch",
        "parameters": [
          {
            "in": "path",
            "name": "task_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Task Id",
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/TaskUpdateRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/TaskUpdateResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "401": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Unauthorized"
          },
          "404": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Not Found"
          },
          "409": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Conflict"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Unprocessable Entity"
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "summary": "Update Task",
        "tags": [
          "tasks"
        ]
      }
    },
...
  }
}